+++ src/helpers/index.js
@@ -36,3 +36,5 @@ export {
   getVocabFromItems,
 } from '@plone/volto/helpers/Vocabularies/Vocabularies';
 export AlignBlock from '@plone/volto/helpers/AlignBlock/AlignBlock';
+
+export Helmet from './Helmet/Helmet';